removed injectable package
moved fixed varaible and injectable varaible into  varaible module
